% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_educ_yrs_column.R
\name{add_educ_yrs_column}
\alias{add_educ_yrs_column}
\title{Add column with recoded education to a CenSoc data table}
\usage{
add_educ_yrs_column(censoc.df)
}
\arguments{
\item{df}{a CenSoc data frame. Must have `EDUC` column.}
}
\value{
a CenSoc data frame with `educ_yrs` column.
}
\description{
Recode education variable.
}
