% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_household_variables.R
\name{add_household_variables}
\alias{add_household_variables}
\title{Match household variables}
\usage{
add_household_variables(censoc, household_census)
}
\arguments{
\item{wcensoc}{a wcensoc file}
}
\value{
Cleaned Wcensoc
}
\description{
Match household variables
}
\keyword{internal}
