% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_race_label_column.R
\name{add_race_label_column}
\alias{add_race_label_column}
\title{Add column with recoded race to a CenSoc data table}
\usage{
add_race_label_column(censoc.df)
}
\arguments{
\item{df}{a CenSoc data frame. Must have `RACE` column.}
}
\value{
a CenSoc data frame with `race_recode` column.
}
\description{
Recode race variable.
}
