% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age_at_death.R
\name{calculate_age_at_death}
\alias{calculate_age_at_death}
\title{Calculate age at Death}
\usage{
calculate_age_at_death(data)
}
\arguments{
\item{data}{data.frame with birth and death info}
}
\value{
data.frame
}
\description{
Calculate age at Death
}
\keyword{internal}
