% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age_at_first_application.R
\name{calculate_age_at_first_app}
\alias{calculate_age_at_first_app}
\title{Date of first application}
\usage{
calculate_age_at_first_app(numapp_condensed = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with number of applications
}
\description{
Date of first application
}
\keyword{internal}
