% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_bunmd_merge.R
\name{census_bunmd_merge}
\alias{census_bunmd_merge}
\title{Merge bunmd and Census}
\usage{
census_bunmd_merge(bunmd = bunmd, census = census,
  census_year = 1940)
}
\arguments{
\item{bunmd}{full bunmd (merged application and death file)}
}
\value{
NUMAPP data.frame with best first name column for each unique ssn (40870456) and cycle dates
}
\description{
Merge bunmd and Census
}
\keyword{internal}
