% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_ss5_merge.R
\name{census_ss5_merge}
\alias{census_ss5_merge}
\title{Merge SS5 and Census}
\usage{
census_ss5_merge(ss5 = ss5, census = census)
}
\arguments{
\item{ss5}{full ss5 (merged application and death file)}
}
\value{
NUMAPP data.frame with best first name column for each unique ssn (40870456) and cycle dates
}
\description{
Merge SS5 and Census
}
\keyword{internal}
