% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_apps.R
\name{clean_apps}
\alias{clean_apps}
\title{Load Application Files}
\usage{
clean_apps(
 
    numapp.file.path = "/censoc/data/numident/1_numident_files_with_original_varnames/applications_original.csv"
)
}
\arguments{
\item{numapp_path}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame
}
\description{
Load Application Files
}
\keyword{internal}
