% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_claims.R
\name{clean_claims}
\alias{clean_claims}
\title{Load claims file}
\usage{
clean_claims(
 
    numclaim.file.path = "/censoc/data/numident/1_numident_files_with_original_varnames/claims_original.csv"
)
}
\arguments{
\item{numapp_path}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame
}
\description{
Load claims file
}
\keyword{internal}
