% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_key}
\alias{clean_key}
\title{Clean matching keys}
\usage{
clean_key(key)
}
\arguments{
\item{key}{a string that acts as a key that may have punctuation. Note that it retains underscores, which are used in the keys.}
}
\value{
a string with no punctuation
}
\description{
Clean matching keys
}
\keyword{internal}
