% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wcensoc.R
\name{clean_wcensoc}
\alias{clean_wcensoc}
\title{Clean WCenSoc}
\usage{
clean_wcensoc(wcensoc, census_year)
}
\arguments{
\item{wcensoc}{a wcensoc file}
}
\value{
Cleaned Wcensoc
}
\description{
Clean WCenSoc
}
\keyword{internal}
