% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_claims.R
\name{condense_claims}
\alias{condense_claims}
\title{Build Numapp File}
\usage{
condense_claims(claims = claims)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame condensed
}
\description{
Build Numapp File
}
\keyword{internal}
