% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_numapp.R
\name{condense_numapp}
\alias{condense_numapp}
\title{Build Numapp File}
\usage{
condense_numapp(numapp = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame condensed
}
\description{
Build Numapp File
}
\keyword{internal}
