% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bunmd.R
\name{create_bunmd}
\alias{create_bunmd}
\title{Unify Numident Files}
\usage{
create_bunmd(claims, deaths, applications)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
Unify Numident Files
}
\description{
Unify Numident Files
}
\keyword{internal}
