% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_censoc_numident.R
\name{create_censoc_numident}
\alias{create_censoc_numident}
\title{Merge bunmd and Census}
\usage{
create_censoc_numident(bunmd = bunmd, census = census, census_year = 1940)
}
\arguments{
\item{bunmd}{full bunmd (merged application and death file)}
}
\value{
NUMAPP data.frame with best first name column for each unique ssn (40870456) and cycle dates
}
\description{
Merge bunmd and Census
}
\keyword{internal}
