% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_bunmd_complete.R
\name{create_weights_bunmd_complete}
\alias{create_weights_bunmd_complete}
\title{Create weights
This function will weight the numident file up to HMD lexis triangles
Restrict data to deaths from 1988 - 2005 with age_at_death between}
\usage{
create_weights_bunmd_complete(
  bunmd.file,
  dyears = c(1988:2005),
  cohorts = c(1895:1940),
  death_ages = c(65:100),
  complete_cases_vars = c("race_last", "bpl")
)
}
\arguments{
\item{data}{data.frame with birth and death info}
}
\value{
data.frame
}
\description{
Create weights
This function will weight the numident file up to HMD lexis triangles
Restrict data to deaths from 1988 - 2005 with age_at_death between
}
\keyword{internal}
