% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_censoc_dmf.R
\name{create_weights_censoc_dmf}
\alias{create_weights_censoc_dmf}
\title{Create weights
This function will weight the numident censoc.dmf up to HMD lexis triangles
Restrict data to certain cohorts and ages at deaths}
\usage{
create_weights_censoc_dmf(
  censoc.dmf,
  cohorts = c(1895:1939),
  death_ages = c(65:100)
)
}
\arguments{
\item{data}{data.frame with birth and death info}
}
\value{
data.frame
}
\description{
Create weights
This function will weight the numident censoc.dmf up to HMD lexis triangles
Restrict data to certain cohorts and ages at deaths
}
\keyword{internal}
