% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_censoc_numident.R
\name{create_weights_censoc_numident}
\alias{create_weights_censoc_numident}
\title{Create weights
This function will weight the numident file up to HMD lexis triangles
Restrict data to deaths from 1988 - 2005 with age_at_death between}
\usage{
create_weights_censoc_numident(
  censoc.numident,
  cohorts = c(1895:1939),
  death_ages = c(65:100),
  hmd_path =
    "/data/josh/CenSoc/hmd/hmd_statistics/deaths/Deaths_lexis/USA.Deaths_lexis.txt"
)
}
\arguments{
\item{data}{data.frame with birth and death info}
}
\value{
data.frame
}
\description{
Create weights
This function will weight the numident file up to HMD lexis triangles
Restrict data to deaths from 1988 - 2005 with age_at_death between
}
\keyword{internal}
