% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_first_word}
\alias{get_first_word}
\title{Get first word from a string}
\usage{
get_first_word(x)
}
\arguments{
\item{x}{a string with spaces between words}
}
\value{
first word from string of words
}
\description{
Get first word from a string
}
\keyword{internal}
