% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_census_ssdm_match.R
\name{load_census}
\alias{load_census}
\title{Load a census file}
\usage{
load_census(census_file, males_only = TRUE)
}
\arguments{
\item{census_file}{string indicating location of file.}

\item{males_only}{whether to just keep males. Default is \code{TRUE}.}

\item{cols_to_keep}{vector indicating colname names of variables to keep. Default is just what is needed for match: first name, last name, age, sex, serial number, person number.}
}
\value{
a census dataframe with match key and unique IPUMS identifier
}
\description{
Load a census file
}
