% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_census_numident_match.R
\name{load_census_numident_match}
\alias{load_census_numident_match}
\title{Load a census file}
\usage{
load_census_numident_match(census_file, cols_to_keep = NULL)
}
\arguments{
\item{census_file}{string indicating location of file.}

\item{cols_to_keep}{vector indicating colname names of variables to keep. Default is just what is needed for match: first name, last name, age, sex, serial number, person number.}

\item{males_only}{whether to just keep males. Default is \code{TRUE}.}
}
\value{
a census dataframe with match key and unique IPUMS identifier
}
\description{
Load a census file
}
\keyword{internal}
