% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dmf_deaths.R
\name{load_dmf_deaths}
\alias{load_dmf_deaths}
\title{Load a census file}
\usage{
load_dmf_deaths(dmf_files)
}
\arguments{
\item{dmf}{a vector of file names where the social security death files live.}
}
\value{
a dataframe with match key, age, and date of death
}
\description{
Load a census file
}
