% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numapp.R
\name{load_numapp}
\alias{load_numapp}
\title{Load Application Files}
\usage{
load_numapp(
 
    numapp.file.path = "/censoc/data/numident/numident_files_with_original_varnames/numident_applications.csv"
)
}
\arguments{
\item{numapp_path}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame
}
\description{
Load Application Files
}
\keyword{internal}
