% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numclaims.R
\name{load_numclaims}
\alias{load_numclaims}
\title{Load claims file}
\usage{
load_numclaims(
 
    numclaim.file.path = "/censoc/data/numident/numident_files_with_original_varnames/numident_claims_complete.csv"
)
}
\arguments{
\item{numapp_path}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame
}
\description{
Load claims file
}
\keyword{internal}
