% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numdeath_2.R
\name{load_numdeath_2}
\alias{load_numdeath_2}
\title{Load numdeath file 2.0}
\usage{
load_numdeath_2(
 
    numdeath.file.path = "/censoc/data/numident/numident_files_with_original_varnames/numident_death_records_complete.csv"
)
}
\arguments{
\item{numdeath_path}{path to the NUMDEATH files}
}
\value{
NUMDEATH data.frame
}
\description{
Load numdeath file 2.0
}
\keyword{internal}
