% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_socsec_deaths.R
\name{load_socsec_deaths}
\alias{load_socsec_deaths}
\title{Load a census file}
\usage{
load_socsec_deaths(socsec_files)
}
\arguments{
\item{socsec_files}{a vector of file names where the social security death files live.}
}
\value{
a dataframe with match key, age, and date of death
}
\description{
Load a census file
}
