% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_numdeath_numapp.R
\name{merge_numapp_numdeath}
\alias{merge_numapp_numdeath}
\title{Merge Application and Death Files}
\usage{
merge_numapp_numdeath(numapp = numapp, numdeath = numdeath, census_year = 1940)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column for each unique ssn (40870456) and cycle dates
}
\description{
Merge Application and Death Files
}
\keyword{internal}
