% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ss5_app_death.R
\name{merge_ss5_app_death}
\alias{merge_ss5_app_death}
\title{Merge SS-5 Application File and SS-5 Death File}
\usage{
merge_ss5_app_death(numapplication = numapp, numdeath = numdeath)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
numapp data.frame with best last name column
}
\description{
Merge SS-5 Application File and SS-5 Death File
}
\keyword{internal}
