% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_app_fwf.R
\name{read_app_fwf}
\alias{read_app_fwf}
\title{Load Application Files from NARA .txt fwf files}
\usage{
read_app_fwf(path = "/data/josh/CenSoc/NUMIAPPLIC/Records/")
}
\arguments{
\item{path}{path to the app file fwf .txt files}
}
\value{
death record data.frame
}
\description{
Load Application Files from NARA .txt fwf files
}
\keyword{internal}
