% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_claim_fwf.R
\name{read_claim_fwf}
\alias{read_claim_fwf}
\title{Load claim files from NARA .txt fwf}
\usage{
read_claim_fwf(path = "/data/josh/CenSoc/NUMICLAIM/Records/")
}
\arguments{
\item{path}{path to the death file fwf .txt files}
}
\value{
death record data.frame
}
\description{
Load claim files from NARA .txt fwf
}
\keyword{internal}
