% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_death_fwf.R
\name{read_death_fwf}
\alias{read_death_fwf}
\title{Load death files from NARA .txt fwf}
\usage{
read_death_fwf(path = "/data/josh/CenSoc/NUMDEATH/")
}
\arguments{
\item{path}{path to the death file fwf .txt files}
}
\value{
death record data.frame
}
\description{
Load death files from NARA .txt fwf
}
\keyword{internal}
