% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_education.R
\name{recode_education}
\alias{recode_education}
\title{recode education
return data.frame with education recoded}
\usage{
recode_education(df, educ_var)
}
\arguments{
\item{data}{data.frame with birth and death info}
}
\value{
data.frame
}
\description{
recode education
return data.frame with education recoded
}
\keyword{internal}
