% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_first_name.R
\name{select_best_first_name}
\alias{select_best_first_name}
\title{Select best first name}
\usage{
select_best_first_name(data = numapp)
}
\arguments{
\item{data}{path to the NUMAPP files}
}
\value{
numapp data.frame with best first name column
}
\description{
Select best first name
}
\keyword{internal}
