% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_birthplace.R
\name{select_birthplace}
\alias{select_birthplace}
\title{Select place of birth}
\usage{
select_birthplace(numapplication = numapp,
  ssn_state_codes = "/home/ipums/andrea-ipums/CenSoc/NUMI_GCBRTHST.txt",
  ssn_country_codes = "/home/ipums/andrea-ipums/CenSoc/nara_bpl.csv")
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select place of birth
}
\keyword{internal}
