% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_birthplace_city.R
\name{select_birthplace_city}
\alias{select_birthplace_city}
\title{Select select Birthplace City}
\usage{
select_birthplace_city(data = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best bpl column and cycle dates
}
\description{
Select select Birthplace City
}
\keyword{internal}
