% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_date_of_first_app.R
\name{select_date_of_first_app}
\alias{select_date_of_first_app}
\title{Date of first application}
\usage{
select_date_of_first_app(numapp = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with number of applications
}
\description{
Date of first application
}
\keyword{internal}
