% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_dob.R
\name{select_dob}
\alias{select_dob}
\title{Select best dob}
\usage{
select_dob(data = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best dob column for each unique ssn (40870456) and cycle dates
}
\description{
Select best dob
}
\keyword{internal}
