% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_longest_last_name.R
\name{select_longest_last_name}
\alias{select_longest_last_name}
\title{Select best father last name}
\usage{
select_longest_last_name(numapp = numapp, variable = "father_fname")
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select best father last name
}
\keyword{internal}
