% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_longest_name.R
\name{select_longest_name}
\alias{select_longest_name}
\title{Select best name (e.g. best first name (fname), best last name (lname))}
\usage{
select_longest_name(data = numapp, name)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select best name (e.g. best first name (fname), best last name (lname))
}
\keyword{internal}
