% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_number_claims.R
\name{select_number_claims}
\alias{select_number_claims}
\title{Calculate total number of claims per individual}
\usage{
select_number_claims(data = claims)
}
\arguments{
\item{numapp}{path to the claims files}
}
\value{
NUMAPP data.frame with number of applications by ssn
}
\description{
Calculate total number of claims per individual
}
\keyword{internal}
