% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_race.R
\name{select_race}
\alias{select_race}
\title{Select race}
\usage{
select_race(numapplication = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select race
}
\keyword{internal}
