% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_race_first.R
\name{select_race_first}
\alias{select_race_first}
\title{Select race}
\usage{
select_race_first(data = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
data.frame with first race for persons with more than one race
}
\description{
Select race
}
\keyword{internal}
