% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_race.R, R/select_race_last.R
\name{select_race_last}
\alias{select_race_last}
\title{Select race}
\usage{
select_race_last(data = numapp)

select_race_last(data = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
data.frame with first and last race

data.frame with last race
}
\description{
Select race

Select race
}
\keyword{internal}
