% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sex.R
\name{select_sex}
\alias{select_sex}
\title{Select sex}
\usage{
select_sex(numapp = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select sex
}
\keyword{internal}
