% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_numident.R
\name{unify_numident}
\alias{unify_numident}
\title{Unify Numident Files}
\usage{
unify_numident(claims_condensed, deaths, apps_condensed)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
Unify Numident Files
}
\description{
Unify Numident Files
}
\keyword{internal}
