% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numapp.R
\name{load_numapp}
\alias{load_numapp}
\title{Load application Files}
\usage{
load_numapp(numapplic_path = "/data/josh/CenSoc/NUMIAPPLIC/Records/")
}
\arguments{
\item{numapp_path}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame
}
\description{
Load application Files
}
\keyword{internal}
