% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numdeath.R
\name{load_numdeath}
\alias{load_numdeath}
\title{Get first word from a string}
\usage{
load_numdeath(numdeath_path = "/data/josh/CenSoc/NUMDEATH/")
}
\arguments{
\item{numdeath_path}{path to the NUMDEATH files}
}
\value{
NUMDEATH data.frame
}
\description{
Get first word from a string
}
\keyword{internal}
