% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_numdeath_geo.R
\name{load_numdeath_geo}
\alias{load_numdeath_geo}
\title{Get first word from a string}
\usage{
load_numdeath_geo(numdeath,
  ssn_state_codes = "/home/ipums/casey-ipums/censoc/data/ssn_state_codes.csv",
  zip_code_states = "/home/ipums/casey-ipums/zip_code_states.csv")
}
\arguments{
\item{numdeath_path}{path to the NUMDEATH files}
}
\value{
NUMDEATH data.frame
}
\description{
Get first word from a string
}
\keyword{internal}
