% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_last_name.R
\name{select_best_last_name}
\alias{select_best_last_name}
\title{Select best last name}
\usage{
select_best_last_name(numapplication = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
numapp data.frame with best last name column
}
\description{
Select best last name
}
\keyword{internal}
