% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_birthplace.R
\name{select_birthplace}
\alias{select_birthplace}
\title{Select place of birth}
\usage{
select_birthplace(numapp = numapp,
  nara_ipums_labels = "/nobackup/90days/andreamg/nara_ipums_labels.csv",
  nara_ipums_state_labels = "/nobackup/90days/andreamg/nara_ipums_state_labels.csv",
  nara_ipums_country_labels = "/nobackup/90days/andreamg/nara_ipums_country_labels.csv")
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best first name column and cycle dates
}
\description{
Select place of birth
}
\keyword{internal}
