% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_birthplace.R
\name{select_birthplace}
\alias{select_birthplace}
\title{Select place of birth}
\usage{
select_birthplace(
  data = numapp,
  crosswalk = "/censoc/data/crosswalks/bpl_crosswalk.csv"
)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with best bpl column and cycle dates
}
\description{
Select place of birth
}
\keyword{internal}
