% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_number_applications.R
\name{select_number_of_apps}
\alias{select_number_of_apps}
\title{Calculate total number of applications per individual}
\usage{
select_number_of_apps(numapp = numapp)
}
\arguments{
\item{numapp}{path to the NUMAPP files}
}
\value{
NUMAPP data.frame with number of applications by ssn
}
\description{
Calculate total number of applications per individual
}
\keyword{internal}
